+++ python/thunder/registration/__init__.py
@@ -1,0 +1,1 @@
+from thunder.registration.methods.crosscorr import CrossCorr, PlanarCrossCorr
