+++ python/thunder/registration/methods/utils.py
